+++ frontend/packages/web/src/hooks/useFormCreateTable.ts
@@ -2,6 +2,7 @@ import { NImage, NImageGroup, NSwitch } from 'naive-ui';
 import dayjs from 'dayjs';
 
 import { PreviewPictureUrl } from '@lib/shared/api/requrls/system/module';
+import { ArchiveStatusEnum, ContractStatusEnum } from '@lib/shared/enums/contractEnum';
 import { FieldTypeEnum, FormDesignKeyEnum } from '@lib/shared/enums/formDesignEnum';
 import { SpecialColumnEnum, TableKeyEnum } from '@lib/shared/enums/tableEnum';
 import { useI18n } from '@lib/shared/hooks/useI18n';
@@ -19,6 +20,7 @@ import {
 } from '@/components/business/crm-form-create/config';
 import type { FormCreateField } from '@/components/business/crm-form-create/types';
 
+import { contractPaymentPlanStatusOptions, contractStatusOptions } from '@/config/contract';
 import { quotationStatusOptions } from '@/config/opportunity';
 import useFormCreateAdvanceFilter from '@/hooks/useFormCreateAdvanceFilter';
 import useReasonConfig from '@/hooks/useReasonConfig';
@@ -686,8 +688,76 @@ export default async function useFormCreateTable(props: FormCreateTableProps) {
         sorter: true,
       },
     ],
-    [FormDesignKeyEnum.CONTRACT]: [], // TODO lmy
-    [FormDesignKeyEnum.CONTRACT_PAYMENT]: [], // TODO lmy
+    [FormDesignKeyEnum.CONTRACT]: [
+      {
+        title: t('org.department'),
+        width: 120,
+        key: 'departmentId',
+        ellipsis: {
+          tooltip: true,
+        },
+        sortOrder: false,
+        sorter: true,
+        render: (row: any) => row.departmentName || '-',
+      },
+      {
+        title: t('contract.status'),
+        width: 120,
+        key: 'status',
+        filterOptions: contractStatusOptions,
+        sortOrder: false,
+        sorter: true,
+        filter: true,
+        render: props.specialRender?.status,
+      },
+      {
+        title: t('contract.voidReason'),
+        width: 120,
+        key: 'voidReason',
+        ellipsis: {
+          tooltip: true,
+        },
+      },
+      {
+        title: t('opportunity.quotation.amount'),
+        width: 120,
+        key: 'amount',
+        sortOrder: false,
+        sorter: true,
+      },
+      {
+        title: t('contract.archivedStatus'),
+        width: 120,
+        key: 'archivedStatus',
+        sortOrder: false,
+        sorter: true,
+        render: (row: any) =>
+          row.archivedStatus === ArchiveStatusEnum.ARCHIVED ? t('common.archive') : t('common.notArchived'),
+      },
+    ],
+    [FormDesignKeyEnum.CONTRACT_PAYMENT]: [
+      {
+        title: t('org.department'),
+        width: 120,
+        key: 'departmentId',
+        ellipsis: {
+          tooltip: true,
+        },
+        sortOrder: false,
+        sorter: true,
+        render: (row: any) => row.departmentName || '-',
+      },
+      {
+        title: t('contract.planStatus'),
+        width: 120,
+        key: 'planStatus',
+        filterOptions: contractPaymentPlanStatusOptions,
+        sortOrder: false,
+        sorter: true,
+        filter: true,
+        render: props.specialRender?.status,
+      },
+    ],
     [FormDesignKeyEnum.PRICE]: [],
   };
   const staticColumns: CrmDataTableColumn[] = [
