+++ frontend/packages/web/src/locale/en-US/common.ts
@@ -202,6 +202,7 @@ export default {
   'common.searchByNamePhone': 'Search by name/phone number',
   'common.searchName': 'Search by name',
   'common.archive': 'archive',
+  'common.notArchived': 'Not archived',
   'common.unarchive': 'Unarchive',
   'common.running': 'Running',
   'common.unExecute': 'Pending',
