+++ frontend/packages/web/src/views/contract/locale/en-US.ts
@@ -4,10 +4,13 @@ export default {
   'contract.my': 'My Contracts',
   'contract.depart': 'Department Contracts',
   'contract.newPlan': 'New plan',
+  'contract.status': 'Contract Status',
+  'contract.archivedStatus': 'Archived Status',
+  'contract.planStatus': 'Plan Status',
+  'contract.voidReason': 'Void Reason',
   'contract.signed': 'Signed',
   'contract.inProgress': 'In Progress',
   'contract.completedPerformance': 'Completed',
-  'contract.void': 'Void',
   'contract.uncompleted': 'Incomplete',
   'contract.partialCompleted': 'Partially Completed',
   'contract.voidedConfirmContent': 'This action cannot be undone. Please proceed with caution!',
