+++ frontend/packages/web/src/views/contract/locale/zh-CN.ts
@@ -4,10 +4,13 @@ export default {
   'contract.my': '我的合同',
   'contract.depart': '部门合同',
   'contract.newPlan': '新建计划',
+  'contract.status': '合同状态',
+  'contract.archivedStatus': '归档状态',
+  'contract.planStatus': '计划状态',
+  'contract.voidReason': '作废原因',
   'contract.signed': '已签署',
   'contract.inProgress': '履行中',
   'contract.completedPerformance': '履行完毕',
-  'contract.void': '作废',
   'contract.uncompleted': '未完成',
   'contract.partialCompleted': '部分完成',
   'contract.voidedConfirmContent': '作废后，不可恢复，请谨慎操作！',
